+++ lib/capybara/node/document_matchers.rb
@@ -7,10 +7,11 @@ module Capybara
       #
       # @!macro title_query_params
       #   @overload $0(string, options = {})
-      #     @param string [String]           The string that title should include
+      #     @param string [String]           The string that title should include      #
       #   @overload $0(regexp, options = {})
       #     @param regexp [Regexp]           The regexp that title should match to
       #   @option options [Numeric] :wait (Capybara.default_max_wait_time) Maximum time that Capybara will wait for title to eq/match given string/regexp argument
+      #   @option options [Boolean] :exact (false) When passed a string should the match be exact or just substring
       # @raise [Capybara::ExpectationNotMet] if the assertion hasn't succeeded during wait time
       # @return [true]
       #